Move C++ WASI configuration to its own header
